+++ js/util.js
@@ -1,0 +1,20 @@
+//https://stackoverflow.com/questions/1527803/generating-random-whole-numbers-in-javascript-in-a-specific-range
+function getRandomInt(min, max) {
+  min = Math.abs(min);
+  max = Math.abs(max);
+
+  if (min < max) {
+    min = Math.ceil(min);       //округление вверх
+    max = Math.floor(max);      //округление вниз
+    return Math.floor(Math.random() * (max - min + 1)) + min;
+  }
+  return min;
+}
+
+function checkLength(checkline, maxLength) {
+  return checkline.length <= maxLength;
+}
+
+checkLength('Какая-то строка', 30);
+
+export {getRandomInt};
